refactor: relocate FormCardType
